Refactor component hierarchy
