+++ app/javascript/components/AdvancedSearch/components/partials/SelectedReference.jsx
@@ -1,9 +1,9 @@
 import React, {useState, useRef, useMemo} from "react";
 import AsyncPaginate from 'react-select-async-paginate';
 import striptags from 'striptags';
-import Translations from '../../Translations/components/Translations';
+import Translations from '../../../Translations/components/Translations';
 
-const SelectedReferenceSearch = (props) => {
+const SelectedReference = (props) => {
   const {
     fieldUuid,
     defaultValues,
@@ -81,4 +81,4 @@ const SelectedReferenceSearch = (props) => {
   );
 };
 
-export default SelectedReferenceSearch;
+export default SelectedReference;
