+++ app/views/advanced_searches/fields/_choice_set_search_field.html.erb
@@ -1,10 +1,10 @@
 <div class="search-container col-lg-12">
   <%=
-    react_component('AdvancedSearchShared/AdvancedContainerSearch',
+    react_component('AdvancedSearch/components/Container',
                     {
                       fieldUuid: field.uuid,
                       defaultValues: value,
-                      childComponent: 'AdvancedChoiceSetSearch/components/ChoiceSetSearch',
+                      childComponent: 'AdvancedSearch/components/ChoiceSet',
                       childProps: {
                         catalog: field.catalog.slug,
                         choiceSet: field.choice_set.attributes.merge(
