+++ app/views/advanced_searches/fields/_date_time_search_field.html.erb
@@ -1,5 +1,5 @@
 <div class="search-container col-lg-12">
-    <%= react_component('AdvancedDateTimeSearch/components/DateTimeSearch',
+    <%= react_component('AdvancedSearch/components/DateTime',
       {
         fieldUuid: field.uuid,
         inputStart: json_hidden_field_id(f, field),
